+++ ExtraBotany-server-1.7.10/src/main/java/com/meteor/extrabotany/common/core/util/SoundHelper.java
@@ -9,7 +9,7 @@ import net.minecraft.world.World;
 
 public class SoundHelper {
     public static void playSound(World world, Sound sound, int x, int y, int z, float fre) {
-        world.playSound(x, y, z, sound.getName(), sound.getVolume(), fre, true);
+        world.playSound((double)x, (double)y, (double)z, sound.getName(), sound.getVolume(), fre, true);
     }
 
     public static void playSoundAtEntity(World world, Sound sound, Entity entity, float fre) {
